+++ src/io/flutter/view/FlutterPerfView.java
@@ -98,7 +98,6 @@ public class FlutterPerfView implements Disposable {
   }
 
   void debugActive(@NotNull FlutterViewMessages.FlutterDebugEvent event) {
-
     final FlutterApp app = event.app;
     final ToolWindowManager toolWindowManager = ToolWindowManager.getInstance(myProject);
     if (!(toolWindowManager instanceof ToolWindowManagerEx)) {
@@ -164,13 +163,21 @@ public class FlutterPerfView implements Disposable {
     final JBRunnerTabs runnerTabs = new JBRunnerTabs(myProject, ActionManager.getInstance(), null, this);
     runnerTabs.setSelectionChangeHandler(this::onTabSelectionChange);
 
-    final List<FlutterDevice> existingDevices = new ArrayList<>();
-    for (FlutterApp otherApp : perAppViewState.keySet()) {
-      existingDevices.add(otherApp.device());
+    final String tabName;
+    final FlutterDevice device = app.device();
+    if (device == null) {
+      tabName = app.getProject().getName();
+    }
+    else {
+      final List<FlutterDevice> existingDevices = new ArrayList<>();
+      for (FlutterApp otherApp : perAppViewState.keySet()) {
+        existingDevices.add(otherApp.device());
+      }
+      tabName = device.getUniqueName(existingDevices);
     }
 
     final JPanel tabContainer = new JPanel(new BorderLayout());
-    final Content content = contentManager.getFactory().createContent(null, app.device().getUniqueName(existingDevices), false);
+    final Content content = contentManager.getFactory().createContent(null, tabName, false);
     tabContainer.add(runnerTabs.getComponent(), BorderLayout.CENTER);
     content.setComponent(tabContainer);
     content.putUserData(ToolWindow.SHOW_CONTENT_ICON, Boolean.TRUE);