+++ testSrc/unit/io/flutter/run/daemon/DaemonEventTest.java
@@ -33,12 +33,12 @@ public class DaemonEventTest {
       // daemon domain
 
       @Override
-      public void onDaemonLogMessage(DaemonEvent.LogMessage event) {
+      public void onDaemonLogMessage(DaemonEvent.DaemonLogMessage event) {
         logEvent(event, event.level, event.message, event.stackTrace);
       }
 
       @Override
-      public void onDaemonShowMessage(DaemonEvent.ShowMessage event) {
+      public void onDaemonShowMessage(DaemonEvent.DaemonShowMessage event) {
         logEvent(event, event.level, event.title, event.message);
       }
 
@@ -123,7 +123,7 @@ public class DaemonEventTest {
   @Test
   public void canReceiveAppStarting() {
     send("app.start", curly("appId:42", "deviceId:456", "directory:somedir", "launchMode:run"));
-    checkLog("AppStarting: 42, 456, somedir", "run");
+    checkLog("AppStarting: 42, 456, somedir, run");
   }
 
   @Test