+++ tool/plugin/lib/plugin.dart
@@ -584,17 +584,7 @@ class BuildCommand extends ProductCommand {
       // TODO: Remove this when we no longer support AS 3.3 (IJ 2018.2.5) or AS 3.4
       var files = <File, String>{};
       var processedFile, source;
-      if (spec.version == '3.3.2') {
-        log('spec.version: ${spec.version}');
-        processedFile = File(
-            'flutter-studio/src/io/flutter/project/FlutterProjectCreator.java');
-        source = processedFile.readAsStringSync();
-        files[processedFile] = source;
-        source = source.replaceAll('List<? extends File>', 'List<File>');
-        processedFile.writeAsStringSync(source);
-      }
-      if (spec.version == '3.3.2' ||
-          spec.version == '3.4') {
+      if (spec.version == '3.4') {
         log('spec.version: ${spec.version}');
         processedFile = File(
             'flutter-studio/src/io/flutter/module/FlutterDescriptionProvider.java');
@@ -617,6 +607,22 @@ class BuildCommand extends ProductCommand {
           'return FlutterIcons.AndroidStudioNewModule;',
           'return IconUtil.toImage(FlutterIcons.AndroidStudioNewModule);',
         );
+        processedFile.writeAsStringSync(source);
+      }
+      if (spec.version != '3.5') {
+        log('spec.version: ${spec.version}');
+        processedFile = File(
+            'flutter-studio/src/io/flutter/project/FlutterProjectModel.java');
+        source = processedFile.readAsStringSync();
+        files[processedFile] = source;
+        source = source.replaceAll('addListener(()', 'addListener(sender');
+        processedFile.writeAsStringSync(source);
+
+        processedFile = File(
+            'flutter-studio/src/io/flutter/project/FlutterSettingsStep.java');
+        source = processedFile.readAsStringSync();
+        files[processedFile] = source;
+        source = source.replaceAll('listen', 'receive');
         processedFile.writeAsStringSync(source);
       }